fix : 어드민페이지 ssr 제거
