+++ src/components/domains/mypage/Admin/AdminPageItem/index.tsx
@@ -1,9 +1,10 @@
 import arrowGary from "@/../public/images/arrowGray.svg";
 import Card from "@/components/commons/Card";
 import AdminFilter from "@/components/commons/DropBottom/AdminFilter";
-import { usePageContext } from "@/pages/mypage/ask";
-import { useFixPageContext } from "@/pages/mypage/fix";
-import { useReportPageContext } from "@/pages/mypage/report";
+import { getAdminBins } from "@/lib/apis/ask";
+import { getAdminBinsFix } from "@/lib/apis/fix";
+import { getAdminBinsReport } from "@/lib/apis/report";
+import { useQuery } from "@tanstack/react-query";
 import classNames from "classnames/bind";
 import Image from "next/image";
 import { useRouter } from "next/router";
@@ -17,25 +18,33 @@ interface AdminProps {
 }
 export default function AdminPageItem({ title }: AdminProps) {
   const [drop, setDrop] = useState<boolean>(false);
-  const [pageFilter, setPageFilter] = useState<string>("전체");
 
   const router = useRouter();
+  const filter = router.query.filter || "전체";
   const binData = (() => {
     switch (router.pathname) {
       case "/mypage/ask":
-        const ask = usePageContext();
-        const askCount = ask.queries[0].state.data.pendingCount;
-        return [askCount, ask.queries[0].state.data.binRegistrationDetails]; // 데이터를 반환
-
+        const { data: ask } = useQuery({
+          queryKey: ["ask", filter],
+          queryFn: () => getAdminBins(filter),
+        });
+        const askCount = ask?.pendingCount;
+        return [askCount, ask?.binRegistrationDetails]; // 데이터를 반환
       case "/mypage/fix":
-        const fix = useFixPageContext();
-        const fixCount = fix.queries[0].state.data.pendingCount;
-        return [fixCount, fix.queries[0].state.data.binModificationDetails]; // 데이터를 반환
+        const { data: fix } = useQuery({
+          queryKey: ["fix", filter],
+          queryFn: () => getAdminBinsFix(filter),
+        });
+        const fixCount = fix?.pendingCount;
+        return [fixCount, fix?.binModificationDetails]; // 데이터를 반환
 
       default:
-        const report = useReportPageContext();
-        const reportCount = report.queries[0].state.data.pendingCount;
-        return [reportCount, report.queries[0].state.data.binComplaintDetails]; // 기본 데이터를 반환
+        const { data: report } = useQuery({
+          queryKey: ["report", filter],
+          queryFn: () => getAdminBinsReport(filter),
+        });
+        const reportCount = report?.pendingCount;
+        return [reportCount, report?.binComplaintDetails]; // 기본 데이터를 반환
     }
   })();
 
@@ -52,13 +61,13 @@ export default function AdminPageItem({ title }: AdminProps) {
             <div className={cn("adminTitleTag")}>{binData[0]}건</div>
           </div>
           <div className={cn("adminDrop")} onClick={handleDrop}>
-            {pageFilter}
+            {filter}
             <Image src={arrowGary} alt="드롭다운" width={10} height={10} />
           </div>
         </div>
 
         <div className={cn("adminCardList")}>
-          {binData[1].map((item: any, index: number) => (
+          {binData[1]?.map((item: any, index: number) => (
             <div
               onClick={() => {
                 router.push(router.route + "/" + item.binId);
@@ -82,13 +91,7 @@ export default function AdminPageItem({ title }: AdminProps) {
         </div>
       </div>
 
-      {drop && (
-        <AdminFilter
-          closeModal={handleDrop}
-          setPageFilter={setPageFilter}
-          pageFilter={pageFilter}
-        />
-      )}
+      {drop && <AdminFilter closeModal={handleDrop} pageFilter={filter} />}
     </>
   );
 }
