+++ src/lib/apis/fix.ts
@@ -1,16 +1,11 @@
 import { instance } from "./axios";
 
-export const getAdminBinsFix = async (data: any, cookies: any) => {
+export const getAdminBinsFix = async (data: any) => {
   const filter =
     data === "전체" ? "ENTIRE" : data === "처리 전" ? "PENDING" : "FINISHED";
   try {
     const res = await instance.get(
-      `/admin/bins/modifications?filter=${filter}`,
-      {
-        headers: {
-          Cookie: cookies || "", // 쿠키를 헤더에 포함하여 API 요청
-        },
-      }
+      `/admin/bins/modifications?filter=${filter}`
     );
     return res.data;
   } catch (err) {
